chore: rename components package
