+++ apps/marketing/package.json
@@ -11,7 +11,7 @@
     "dev": "next dev"
   },
   "dependencies": {
-    "@deepdish/components": "*",
+    "@deepdish/ui": "*",
     "next": "14.2.4",
     "react": "18.3.1",
     "react-dom": "18.3.1"
