+++ packages/ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@deepdish/components",
+  "name": "@deepdish/ui",
   "version": "0.0.0",
   "description": "DeepDish UI components",
   "homepage": "",
